requirements.txt
autopep8==1.5.4
beautifulsoup4==4.9.3
boto3==1.16.47
botocore==1.19.47
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
jmespath==0.10.0
more-itertools==8.6.0
numpy==1.19.4
pandas==1.2.0
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
soupsieve==2.1
toml==0.10.2
urllib3==1.26.2
